Refactor measurement API and autentication models
The commit implements a struct for all measurement API arguments to clean up the function call. The struct includes debug implementation and is seralizable. It has also renamed the 'auth' model fields for clarity and improved some of its comments. All references to the old 'auth' models are also updated in the code.